+++ docs/source/release_notes.rst
@@ -10,6 +10,8 @@ Release notes
 
 - Added Global [Lee's statistic](https://onlinelibrary.wiley.com/doi/abs/10.1111/gean.12106), along with a note on weighted product that upon z-scaling it is equivalent to Lee's local statistic.
 
+- The Global L statistic and Global [Moran's R](https://www.nature.com/articles/s41467-023-39608-w) are themselves basically identical.
+
 - Changed the ``li.mt.bivar`` parameter ``function_name`` to ``local_name`` for consistency and to avoid ambiguity with the newly-added ``global_name`` parameter.
