+++ liana/method/sp/_bivariate/_global_functions.py
