+++ liana/method/sp/_bivariate/_local_functions.py
