+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "liana"
-version = "1.0.5"
+version = "1.0.6"
 description = "LIANA+: a one-stop-shop framework for cell-cell communication"
 authors = ["Daniel Dimitrov <daniel.dimitrov@uni-heidelberg.de>"]
 classifiers = [
